backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
keras
tf-keras
transformers
pillow
numpy
python-multipart
gdown
requests